requirements.txt
boto3
aws_cdk.core
aws_cdk.aws_cloudformation
aws_cdk.aws_dynamodb
aws_cdk.aws_lambda
aws_cdk.aws_lambda_event_sources
aws_cdk.aws_apigateway
aws_cdk.aws_stepfunctions
aws_cdk.aws_stepfunctions_tasks
aws_cdk.aws_iam
aws_cdk.aws_ssm
-e .
